+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## next
+## 1.0.0-beta.29 (24-02-2020)
 
 - Fixed `tree` view to apply nested `itemConfig` when `limitLines` is used
 - Added `'children'` as default query for children in `tree` view